Type system and validation overhaul
The type system was very flexible but had no constraints. Generics
and vararg were validated with custom validation implementations
in the different methods. This made it hard to implement generics
and vararg for user functions. I also had plans to add higher order
functions but this got difficult as functions like `arrayList` had
lots of accepted argument formats while only being one function.
A higher order function of `arrayList` did therefore not have a
proper type.

The new system only has one validation method and two more types
have been added: generic and vararg. This change does not add any
features but paves the way for:

1. Higher order functions
2. Vararg in user functions
3. generics in user functions

I also removed the need for types in runtime. It was only needed
in a few places that which easy to get around. This means that quite
a few bytes can be saved in the compiled program.